web/requirements.txt
Django==1.11.3
djangorestframework==3.6.3
mysqlclient==1.3.10